gruvin9x issue 81

+++ pers.cpp
@@ -283,7 +283,7 @@ void eeLoadModel(uint8_t id)
 
 #ifdef SIMU
     if (sz > 0 && sz != sizeof(g_model)) {
-      printf("Model data read=%d bytes vs %lu bytes\n", sz, sizeof(ModelData));
+      printf("Model data read=%d bytes vs %d bytes\n", sz, (int)sizeof(ModelData));
     }
 #endif
